+++ packages/radix-ui-themes/src/components/card.tsx
@@ -35,8 +35,7 @@ const Card = React.forwardRef<CardElement, CardProps>((props, forwardedRef) => {
       ref={forwardedRef}
       {...cardProps}
       className={classNames(
-        'rt-reset-a',
-        'rt-reset-button',
+        'rt-reset',
         'rt-Card',
         className,
         withBreakpoints(size, 'rt-r-size'),